requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
aiosqlite==0.21.0
altair==5.5.0
altgraph==0.17.4
annotated-types==0.7.0
anthropic==0.69.0
anyio==4.11.0
asttokens==3.0.0
asyncer==0.0.9
attrs==25.4.0
audioop-lts==0.2.2
backoff==2.2.1
banks==2.2.0
beautifulsoup4==4.14.2
bidict==0.23.1
blinker==1.9.0
bm25s==0.2.14
Brotli==1.1.0
cachetools==6.2.0
certifi==2025.10.5
chainlit==2.8.3
charset-normalizer==3.4.3
chevron==0.14.0
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cuid==0.4
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
docstring_parser==0.17.0
dotenv==0.9.9
executing==2.2.1
fastapi==0.116.2
ffmpy==0.6.3
filelock==3.20.0
filetype==1.2.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.41.1
google-genai==1.42.0
googleapis-common-protos==1.70.0
gradio_client==1.13.3
greenlet==3.2.4
griffe==1.14.0
groovy==0.1.2
grpcio==1.75.1
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.35.3
idna==3.10
importlib_metadata==8.7.0
inflection==0.5.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonpickle==4.1.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
Lazify==0.4.0
literalai==0.1.201
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.14.4
llama-index-cli==0.5.3
llama-index-core==0.14.4
llama-index-embeddings-google-genai==0.3.1
llama-index-embeddings-openai==0.5.1
llama-index-indices-managed-llama-cloud==0.9.4
llama-index-instrumentation==0.4.1
llama-index-llms-google-genai==0.6.2
llama-index-llms-openai==0.6.4
llama-index-readers-file==0.5.4
llama-index-readers-llama-parse==0.5.1
llama-index-retrievers-bm25==0.6.5
llama-index-utils-workflow==0.4.1
llama-index-workflows==2.7.1
llama-parse==0.6.54
macholib==1.16.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.7
matplotlib-inline==0.1.7
mcp==1.17.0
mdurl==0.1.2
monotonic==1.6
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.7.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.2
numpy==2.3.3
openai==1.109.1
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-alephalpha==0.47.3
opentelemetry-instrumentation-anthropic==0.47.3
opentelemetry-instrumentation-bedrock==0.47.3
opentelemetry-instrumentation-chromadb==0.47.3
opentelemetry-instrumentation-cohere==0.47.3
opentelemetry-instrumentation-crewai==0.47.3
opentelemetry-instrumentation-google-generativeai==0.47.3
opentelemetry-instrumentation-groq==0.47.3
opentelemetry-instrumentation-haystack==0.47.3
opentelemetry-instrumentation-lancedb==0.47.3
opentelemetry-instrumentation-langchain==0.47.3
opentelemetry-instrumentation-llamaindex==0.47.3
opentelemetry-instrumentation-logging==0.58b0
opentelemetry-instrumentation-marqo==0.47.3
opentelemetry-instrumentation-mcp==0.47.3
opentelemetry-instrumentation-milvus==0.47.3
opentelemetry-instrumentation-mistralai==0.47.3
opentelemetry-instrumentation-ollama==0.47.3
opentelemetry-instrumentation-openai==0.47.3
opentelemetry-instrumentation-openai-agents==0.47.3
opentelemetry-instrumentation-pinecone==0.47.3
opentelemetry-instrumentation-qdrant==0.47.3
opentelemetry-instrumentation-redis==0.58b0
opentelemetry-instrumentation-replicate==0.47.3
opentelemetry-instrumentation-requests==0.58b0
opentelemetry-instrumentation-sagemaker==0.47.3
opentelemetry-instrumentation-sqlalchemy==0.58b0
opentelemetry-instrumentation-threading==0.58b0
opentelemetry-instrumentation-together==0.47.3
opentelemetry-instrumentation-transformers==0.47.3
opentelemetry-instrumentation-urllib3==0.58b0
opentelemetry-instrumentation-vertexai==0.47.3
opentelemetry-instrumentation-watsonx==0.47.3
opentelemetry-instrumentation-weaviate==0.47.3
opentelemetry-instrumentation-writer==0.47.3
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-semantic-conventions-ai==0.4.13
opentelemetry-util-http==0.58b0
orjson==3.11.3
packaging==25.0
pandas==2.2.3
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.5.0
posthog==3.25.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.32.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.10
pydantic-settings==2.11.0
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.2
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
PyJWT==2.10.1
pyparsing==3.2.5
pypdf==6.1.1
PyStemmer==2.2.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.3
python-multipart==0.0.20
python-socketio==5.14.1
pytz==2025.2
pyvis==0.3.2
PyYAML==6.0.3
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
rich==14.2.0
rpds-py==0.27.1
rsa==4.9.1
ruff==0.14.0
safehttpx==0.1.6
scipy==1.16.2
semantic-version==2.10.0
setuptools==80.9.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.48.0
streamlit==1.50.0
striprtf==0.0.26
syncer==2.0.3
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.1
toml==0.10.2
tomli==2.3.0
tomlkit==0.13.3
tornado==6.5.2
tqdm==4.67.1
traceloop-sdk==0.47.3
traitlets==5.14.3
typer==0.19.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
watchfiles==0.24.0
wcwidth==0.2.14
websockets==15.0.1
wrapt==1.17.3
wsproto==1.2.0
yarl==1.22.0
zipp==3.23.0
